+++ packages/ipfs-http-client/src/block/get.js
@@ -16,7 +16,8 @@ module.exports = configure(api => {
       searchParams: toUrlSearchParams({
         arg: cid.toString(),
         ...options
-      })
+      }),
+      headers: options.headers
     })
 
     return new Block(Buffer.from(await res.arrayBuffer()), cid)