+++ packages/ipfs-http-client/src/key/rm.js
@@ -12,7 +12,8 @@ module.exports = configure(api => {
       searchParams: toUrlSearchParams({
         arg: name,
         ...options
-      })
+      }),
+      headers: options.headers
     })
     const data = await res.json()