+++ packages/ipfs-http-client/src/lib/multipart-request.js
@@ -5,8 +5,9 @@ const toStream = require('./to-stream')
 const { nanoid } = require('nanoid')
 const modeToString = require('../lib/mode-to-string')
 const mtimeToObject = require('../lib/mtime-to-object')
+const merge = require('merge-options').bind({ ignoreUndefined: true })
 
-async function multipartRequest (source, abortController, boundary = `-----------------------------${nanoid()}`) {
+async function multipartRequest (source, abortController, headers = {}, boundary = `-----------------------------${nanoid()}`) {
   async function * streamFiles (source) {
     try {
       let index = 0
@@ -58,9 +59,9 @@ async function multipartRequest (source, abortController, boundary = `----------
   }
 
   return {
-    headers: {
+    headers: merge(headers, {
       'Content-Type': `multipart/form-data; boundary=${boundary}`
-    },
+    }),
     body: await toStream(streamFiles(source))
   }
 }