+++ packages/ipfs-http-client/src/object/patch/set-data.js
@@ -24,7 +24,7 @@ module.exports = configure(api => {
         ...options
       }),
       ...(
-        await multipartRequest(data, controller)
+        await multipartRequest(data, controller, options.headers)
       )
     })).json()