+++ packages/ipfs-http-client/src/object/stat.js
@@ -13,7 +13,8 @@ module.exports = configure(api => {
       searchParams: toUrlSearchParams({
         arg: `${Buffer.isBuffer(cid) ? new CID(cid) : cid}`,
         ...options
-      })
+      }),
+      headers: options.headers
     })
 
     return res.json()