+++ packages/ipfs-http-client/src/pin/rm.js
@@ -12,7 +12,8 @@ module.exports = configure(api => {
       searchParams: toUrlSearchParams({
         arg: `${path}`,
         ...options
-      })
+      }),
+      headers: options.headers
     })
 
     const data = await res.json()